moving remotelockstore to stores class
